Add gometalinter

+++ ratecounter.go
@@ -27,6 +27,7 @@ func NewRateCounter(intrvl time.Duration) *RateCounter {
 	return ratecounter.WithResolution(20)
 }
 
+// WithResolution determines the minimum resolution of this counter, default is 20
 func (r *RateCounter) WithResolution(resolution int) *RateCounter {
 	if resolution < 1 {
 		panic("RateCounter resolution cannot be less than 1")
@@ -52,7 +53,7 @@ func (r *RateCounter) run() {
 			next := (int(current) + 1) % r.resolution
 			r.counter.Incr(-1 * r.partials[next].Value())
 			r.partials[next].Reset()
-			atomic.CompareAndSwapInt32(&r.current, int32(current), int32(next))
+			atomic.CompareAndSwapInt32(&r.current, current, int32(next))
 			if r.counter.Value() == 0 {
 				atomic.StoreInt32(&r.running, 0)
 				ticker.Stop()
